chore(sdk): move back action proxy in dedicated folder
