+++ packages/sdk/src/plugin/action-proxy/types.test.ts
@@ -1,8 +1,8 @@
 import { test } from 'vitest'
-import { EmptyPlugin, FooBarBazPlugin } from '../../../fixtures'
+import { EmptyPlugin, FooBarBazPlugin } from '../../fixtures'
 import { ActionProxy } from './types'
-import * as utils from '../../../utils/type-utils'
-import { BasePlugin } from '../../common'
+import * as utils from '../../utils/type-utils'
+import { BasePlugin } from '../common'
 
 test('ActionProxy of FooBarBazPlugin should reflect actions of bot, integration and interface deps', async () => {
   type Actual = ActionProxy<FooBarBazPlugin>
