+++ packages/sdk/src/plugin/implementation.ts
@@ -11,9 +11,9 @@ import type {
 } from '../bot'
 import { WorkflowProxy, proxyWorkflows } from '../bot/workflow-proxy'
 import * as utils from '../utils'
+import { ActionProxy, proxyActions } from './action-proxy'
 import { BasePlugin, PluginConfiguration, PluginInterfaceExtensions, PluginRuntimeProps } from './common'
 import { formatEventRef, parseEventRef, resolveEvent } from './interface-resolution'
-import * as proxies from './proxies'
 import {
   ActionHandlers,
   MessageHandlers,
@@ -39,6 +39,7 @@ import {
   WorkflowPayloads,
   HookDefinitionType,
 } from './server/types'
+import { proxyStates, StateProxy } from './state-proxy'
 
 export type PluginImplementationProps<TPlugin extends BasePlugin = BasePlugin> = {
   actions: ActionHandlers<TPlugin>
@@ -47,8 +48,8 @@ export type PluginImplementationProps<TPlugin extends BasePlugin = BasePlugin> =
 type Tools<TPlugin extends BasePlugin = BasePlugin> = {
   configuration: PluginConfiguration<TPlugin>
   interfaces: PluginInterfaceExtensions<TPlugin>
-  actions: proxies.action.ActionProxy<TPlugin>
-  states: proxies.state.StateProxy<TPlugin>
+  actions: ActionProxy<TPlugin>
+  states: StateProxy<TPlugin>
   workflows: WorkflowProxy<TPlugin>
   alias?: string
 }
@@ -98,8 +99,8 @@ export class PluginImplementation<TPlugin extends BasePlugin = BasePlugin> imple
 
   private _getTools(client: BotSpecificClient<any>): Tools {
     const { configuration, interfaces, alias } = this._runtime
-    const actions = proxies.action.proxyActions(client, this._runtime) as proxies.action.ActionProxy<BasePlugin>
-    const states = proxies.state.proxyStates(client, this._runtime) as proxies.state.StateProxy<BasePlugin>
+    const actions = proxyActions(client, this._runtime) as ActionProxy<BasePlugin>
+    const states = proxyStates(client, this._runtime) as StateProxy<BasePlugin>
     const workflows = proxyWorkflows(client) as WorkflowProxy<BasePlugin>
 
     return {
