--- packages/sdk/src/plugin/proxies/index.ts
@@ -1,2 +1,0 @@
-export * as action from './action'
-export * as state from './state'
