+++ packages/sdk/src/plugin/server/types.ts
@@ -2,8 +2,9 @@ import * as client from '@botpress/client'
 import * as bot from '../../bot'
 import * as workflowProxy from '../../bot/workflow-proxy'
 import * as utils from '../../utils/type-utils'
+import * as actionProxy from '../action-proxy'
 import * as common from '../common'
-import * as proxies from '../proxies'
+import * as stateProxy from '../state-proxy'
 
 type EnumeratePluginEvents<TPlugin extends common.BasePlugin> = bot.EnumerateEvents<TPlugin> &
   common.EnumerateInterfaceEvents<TPlugin>
@@ -99,7 +100,8 @@ export type CommonHandlerProps<TPlugin extends common.BasePlugin> = {
   client: PluginClient<TPlugin>
   configuration: common.PluginConfiguration<TPlugin>
   interfaces: common.PluginInterfaceExtensions<TPlugin>
-  actions: proxies.action.ActionProxy<TPlugin>
+  actions: actionProxy.ActionProxy<TPlugin>
+  stateProxy: stateProxy.StateProxy<TPlugin>
 
   /**
    * # EXPERIMENTAL
